pyproject.toml
[project]
name = "Langchain_training_project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=0.3.4",
    "langchain>=0.3.18",
    "openai>=1.61.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
    "langchain-community>=0.3.17",
    "chromadb>=0.6.3",
    "pypdf>=5.3.0",
    "cryptography>=44.0.1",
    "langchain-chroma>=0.2.2",
]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
lint.select = ["E", "F", "B", "I", "SIM",]
exclude = ["settings.py","migrations"]
line-length = 79

[tool.mypy]
exclude= []
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
disable_error_code = ["import-untyped"]
